requirements.txt
python-dotenv
pymysql
openai
requests
Flask